+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - verify method improvements #97
 
+### Changed
+- make parents dat url validation strict #96
+
+
 ## 0.5.2 - 2020-02-28
 ### Added
 - new changelog file