+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - make parents dat url validation strict #96
+- switch to GitHub actions for CI #114
 
 ### Fixed
 - leveldb open errors ignored #112