+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.3] - 2020-03-10
 ### Added
 - verify method improvements #97
 - update to module specs 0.2.3 #99
@@ -26,4 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ready method call is now implicit
 - extra params now throw validationError
 
-[Unreleased]: https://github.com/p2pcommons/sdk-js/compare/v0.5.2...HEAD
+[Unreleased]: https://github.com/p2pcommons/sdk-js/compare/v0.5.3...HEAD
+[0.5.3]: https://github.com/p2pcommons/sdk-js/compare/v0.5.2...v0.5.3